requirements.txt
nvidia-ml-py3==7.352.0
PyYAML==6.0.1
requests==2.31.0
python-telegram-bot==20.7